+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 24-03-2025
+
+### Added
+- Breakpoints parser - '200px' to 200
+- alignments & border
+
 ## [0.1.0] - 24-03-2025
 
 ### Added